Fix coveragerc and rename executor
